tests/experiments/cost_scaling/requirements.txt
dspy-ai>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
scikit-learn>=1.0.0
faiss-cpu>=1.7.0
scipy>=1.7.0
networkx>=2.6.0 